Temporary hack for dual core cpus
The 2nd sometimes doesnt work, this would make the phone hang in boot,
we should restart the phone to restart everything.

This is a temporary hack to solve this problem, phone will auto-reboot on errors.

(edit t.p.)
taskset or a similar methode should then be used to set core affinity to the
same core as init process, its a ptrace() limit

Change-Id: Iada343945634194b5e583adcae24fcfa57b4b56b